requirements.txt
alembic==1.4.3
aniso8601==7.0.0
appdirs==1.4.4
black==20.8b1
certifi==2020.6.20
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
databases==0.3.2
distlib==0.3.1
filelock==3.0.12
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
identify==1.5.5
idna==2.10
importlib-metadata==2.0.0
Mako==1.1.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
nodeenv==1.5.0
pathspec==0.8.0
pre-commit==2.7.1
promise==2.3
pydantic==1.6.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
Rx==1.6.1
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.8
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
virtualenv==20.0.31
websockets==8.1
zipp==3.2.0
